Added graphql subscription in workflow table
